[meta,refactor,doc,style]: remove mentions of DBL, use non-scoped namespaces, remove redundant summary docstring indentations, prettier
