--- DiscordBotsList.Webhooks/Data/PartialProject.cs
@@ -1,37 +1,0 @@
-using System.Text.Json.Serialization;
-using DiscordBotsList.Webhooks.Serialization;
-
-namespace DiscordBotsList.Webhooks.Data
-{
-    /// <summary>
-    ///     A brief information on project listed on Top.gg.
-    /// </summary>
-    public class PartialProject
-    {
-        /// <summary>
-        ///     The project's ID.
-        /// </summary>
-        [JsonPropertyName("id")]
-        [JsonConverter(typeof(ULongToStringConverter))]
-        public ulong Id { get; internal init; }
-
-        /// <summary>
-        ///     The project's ID.
-        /// </summary>
-        [JsonPropertyName("type")]
-        public ProjectType Type { get; internal init; }
-
-        /// <summary>
-        ///     The project's platform.
-        /// </summary>
-        [JsonPropertyName("platform")]
-        public Platform Platform { get; internal init; }
-
-        /// <summary>
-        ///     The project's platform ID.
-        /// </summary>
-        [JsonPropertyName("platform_id")]
-        [JsonConverter(typeof(ULongToStringConverter))]
-        public ulong PlatformId { get; internal init; }
-    }
-}
