--- DiscordBotsList.Webhooks/DiscordBotsList.Webhooks.sln
@@ -1,24 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.2.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiscordBotsList.Webhooks", "DiscordBotsList.Webhooks.csproj", "{1D925476-7407-E766-5044-48D0FCF35938}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1D925476-7407-E766-5044-48D0FCF35938}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D925476-7407-E766-5044-48D0FCF35938}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D925476-7407-E766-5044-48D0FCF35938}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D925476-7407-E766-5044-48D0FCF35938}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2888D49C-0344-4012-8648-7B2959A50321}
-	EndGlobalSection
-EndGlobal
