+++ Topgg.Sdk.Webhooks/Tests/Mocks/VoteCreate.json
