+++ Topgg.Sdk.Webhooks/Topgg.Sdk.Webhooks.csproj
@@ -3,7 +3,7 @@
         <TargetFramework>net8.0</TargetFramework>
         <Authors>Velddev, Faith, null8626</Authors>
         <Company>Top.gg</Company>
-        <Description>The community-maintained ASP.NET webhooks library for Top.gg.</Description>
+        <Description>The community-maintained .NET SDK for Top.gg Webhooks.</Description>
         <Copyright>Mike Veldsink</Copyright>
         <PackageProjectUrl>https://github.com/Top-gg-Community/dotnet-sdk</PackageProjectUrl>
         <RepositoryUrl>https://github.com/Top-gg-Community/dotnet-sdk</RepositoryUrl>
